+++ src/elements/video-360.js
@@ -7,22 +7,19 @@ import {DeviceOrientationControls} from 'three/examples/jsm/controls/DeviceOrien
 import {OrbitControls} from 'three/examples/jsm/controls/OrbitControls';
 import {
   DeviceOrientationHelper
-} from '../utils/deviceorientation-helper';import {VRHelper} from '../utils/vr-helper';
+} from '../utils/deviceorientation-helper';import {XRHelper} from '../utils/xr-helper';
 import {THREEHelper} from '../utils/three-helper';
 
 class Video360 extends HTMLElement {
   connectedCallback() {
     Promise.all([
-      VRHelper.getVRDevice(),
       DeviceOrientationHelper.hasDeviceOrientation()
     ]).then(array => {
       this._initialize(array[0], array[1]);
     });
   }
 
-  _initialize(device, hasDeviceOrientation) {
-    const hasDevice = device !== null;
-
+  _initialize(hasDeviceOrientation) {
     // Attributes
 
     let src = this.getAttribute('src') || '';
@@ -125,16 +122,13 @@ class Video360 extends HTMLElement {
 
     // VR / Fullscreen
 
-    const button = VRHelper.createButton(renderer.domElement, device);
+    const button = XRHelper.createButton(renderer, controls);
     container.appendChild(button);
 
-    THREEHelper.setupVRModeSwitching(renderer, camera, controls, device);
-
-    window.addEventListener('vrdisplaypresentchange', event => {
+    renderer.vr.addEventListener('sessionstart', event => {
       triggered = true;
       play();
-    }, false);
-
+    });
     document.addEventListener('fullscreenchange', event => {
       triggered = true;
       play();
@@ -159,7 +153,7 @@ class Video360 extends HTMLElement {
 
         renderer.setSize(width, height);
 
-        VRHelper.updateButton(renderer.domElement, button);
+        XRHelper.updateButton(renderer.domElement, button);
       }
 
       if (newLoop !== loop) {
